+++ src/collective/eeafaceted/z3ctable/columns.py
@@ -782,7 +782,7 @@ class ActionsColumn(BrowserViewCallColumn):
     """
 
     header_js = '<script type="text/javascript">jQuery(document).ready(initializeOverlays);' \
-                'jQuery(document).ready(preventDefaultClickTransition);</script>'
+                'jQuery(document).ready(preventDefaultClick);</script>'
     view_name = 'actions_panel'
     params = {'showHistory': True, 'showActions': True}
     # not necessary to escape, everything is generated